Apply problem-solving methodology
Created a new markdown file 'problem-solving.md'. Included a detailed
description of the problem solving methodology covered as part of The
Odin Project Foundations Course. Further included a detailed description
of this process being applied to this particular project.

Implemented pseudocode, as comments, in the JavaScript file
'rock-paper-scissors.js' that was created as part of the problem-solving
process in 'problem-solving.md'

Updated the README file to correct typos and spelling errors.